Resultados

+++ fsecuential.c
@@ -10,17 +10,17 @@
 #define TRUE 1
 #define FALSE 0
 
-void imprimematriz(int *matriz, int n);
-void MatrixMultiply(int n, int *a, int *b, int *c);
-int *create_array_as_matrix(int r, int c);
-void populate_array_as_matrix(int *arr, int r, int c);
-int array_as_matrix_equals(int *a, int *b, int r, int c);
+void imprimematriz(double *matriz, int n);
+void MatrixMultiply(int n, double *a, double *b, double *c);
+double *create_array_as_matrix(int r, int c);
+void populate_array_as_matrix(double *arr, int r, int c);
+int array_as_matrix_equals(double *a, double *b, int r, int c);
 
-int *a;
-int *b;
-int *c;
+double *a;
+double *b;
+double *c;
 
-void MatrixMultiply(int n, int *a, int *b, int *c)
+void MatrixMultiply(int n, double *a, double *b, double *c)
 {
     int i, j, k;
     for (i = 0; i < n; i++)
@@ -29,7 +29,7 @@ void MatrixMultiply(int n, int *a, int *b, int *c)
                 c[i * n + j] += a[i * n + k] * b[k * n + j];
 }
 
-void imprimematriz(int *matriz, int n)
+void imprimematriz(double *matriz, int n)
 {
     int i;
     int j;
@@ -37,7 +37,7 @@ void imprimematriz(int *matriz, int n)
     {
         for (j = 0; j < n; j++)
         {
-            printf("%d\t", matriz[i * n + j]);
+            printf("%0.2f\t", matriz[i * n + j]);
         }
         printf("\n");
     }
@@ -69,8 +69,7 @@ int main(int argc, char *argv[])
 
     if (mi_rank == 0)
     {
-
-        /* -----SECUENCIAL------ */
+         /* -----SECUENCIAL------ */
 
         /* Mide el tiempo del algoritmo secuencial */
         gettimeofday(&start2, 0);
@@ -79,7 +78,7 @@ int main(int argc, char *argv[])
         // imprimematriz(&c[0], N);
         gettimeofday(&stop, 0);
         // Se detiene el tiempo y se muestran los resultados
-        fprintf(stdout, "%0.10f\n",
+        fprintf(stdout, "%0.8f\n",
                 (stop.tv_sec + stop.tv_usec * 1e-6) - (start2.tv_sec + start2.tv_usec * 1e-6));
     }
 
@@ -87,13 +86,13 @@ int main(int argc, char *argv[])
     return 0;
 }
 
-int *create_array_as_matrix(int r, int c)
+double *create_array_as_matrix(int r, int c)
 {
-    int *mat = calloc(r * c, sizeof(int));
+    double *mat = calloc(r * c, sizeof(double));
     return mat;
 }
 
-void populate_array_as_matrix(int *arr, int r, int c)
+void populate_array_as_matrix(double *arr, int r, int c)
 {
     int j;
     for (j = 0; j < r * c; j++)
@@ -102,7 +101,7 @@ void populate_array_as_matrix(int *arr, int r, int c)
     }
 }
 
-int array_as_matrix_equals(int *a, int *b, int r, int c)
+int array_as_matrix_equals(double *a, double *b, int r, int c)
 {
     int i = 0;
     for (i = 0; i < r * c; i++)
